pyproject.toml
[tool.poetry]
name = "mo2-download-manager"
version = "1.2.0"
description = "Adds a few new features to manage downloads within MO2"
authors = ["aglowinthefield"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.8.0,<3.13"
pyqt5-stubs = "^5.14.2"
pyqt6 = { version = "6.8.0", python = ">=3.9" }
mobase-stubs = [
    { version = "<=2.4.0", python = ">=3.8.0,<3.9.dev0" },
    { version = "2.5.2", python = ">=3.12" }
]
pylint = [
    { version = "^3.3.2", python = ">=3.9" },
    { version = "3.2.7", python = "< 3.9" }
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
